pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "testrail-mcp"
version = "0.1.0"
description = "TestRail MCP Server"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "python-mcp>=1.0.0",
    "fastmcp",
]

[project.scripts]
testrail-mcp = "testrail_mcp.main:main"

[tool.hatch.build.targets.wheel]
packages = ["testrail_mcp"]